consumer/requirements.txt
kafka-python==2.0.2
loguru==0.6.0
psycopg2==2.9.3
sqlalchemy==1.4.37
